+++ bluepysnap/morph.py
@@ -20,8 +20,8 @@
 from pathlib import Path
 
 import morph_tool.transform as transformations
+import morphio
 import numpy as np
-from morphio.mut import Morphology
 
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.sonata_constants import Node
@@ -51,8 +51,8 @@ class MorphHelper:
         self._alternate_morphologies = alternate_morphologies or {}
         self._population = population
 
-    def get_morphology_dir(self, extension):
-        """Return morphology directory based on a given extension."""
+    def _get_morphology_base(self, extension):
+        """Get morphology base path; this will be a directory unless it's a morphology container."""
         if extension == "swc":
             if not self._morph_dir:
                 raise BluepySnapError("'morphologies_dir' is not defined in config")
@@ -68,6 +68,25 @@ class MorphHelper:
 
         return morph_dir
 
+    def get_morphology_dir(self, extension="swc"):
+        """Return morphology directory based on a given extension."""
+        morph_dir = self._get_morphology_base(extension)
+
+        if extension == "h5" and Path(morph_dir).is_file():
+            raise BluepySnapError(
+                f"'{morph_dir}' is a morphology container, so a directory does not exist"
+            )
+
+        return morph_dir
+
+    def get_name(self, node_id):
+        """Get the morphology name for a `node_id`."""
+        if not is_node_id(node_id):
+            raise BluepySnapError("node_id must be a int or a CircuitNodeId")
+
+        name = self._population.get(node_id, Node.MORPHOLOGY)
+        return name
+
     def get_filepath(self, node_id, extension="swc"):
         """Return path to SWC morphology file corresponding to `node_id`.
 
@@ -75,9 +94,7 @@ class MorphHelper:
             node_id (int/CircuitNodeId): could be a int or CircuitNodeId.
             extension (str): expected filetype extension of the morph file.
         """
-        if not is_node_id(node_id):
-            raise BluepySnapError("node_id must be a int or a CircuitNodeId")
-        name = self._population.get(node_id, Node.MORPHOLOGY)
+        name = self.get_name(node_id)
 
         return Path(self.get_morphology_dir(extension), f"{name}.{extension}")
 
@@ -90,11 +107,19 @@ class MorphHelper:
                 according to `node_id` position in the circuit.
             extension (str): expected filetype extension of the morph file.
         """
-        filepath = self.get_filepath(node_id, extension=extension)
-        result = Morphology(filepath)
+        collection = morphio.Collection(
+            self._get_morphology_base(extension),
+            [
+                f".{extension}",
+            ],
+        )
+        name = self.get_name(node_id)
+        result = collection.load(name, mutable=True)
+
         if transform:
             T = np.eye(4)
             T[:3, :3] = self._population.orientations(node_id)  # rotations
             T[:3, 3] = self._population.positions(node_id).values  # translations
             transformations.transform(result, T)
+
         return result.as_immutable()